fix for duplicate issues
